move to_bytes to utils mod
